requirements.txt
asgiref==3.7.2
Django==5.0
django-filter==23.5
sqlparse==0.4.4